requirements.txt
numpy==1.26.2
Pillow==10.1.0
PyYAML==6.0.1
scipy==1.11.4
termcolor==2.3.0
torch==1.13.0+cu116
torchvision==0.14.0+cu116
